chore: fix broken localization initialization
