+++ template/package.json
@@ -36,7 +36,7 @@
     "lottie-react-native": "5.1.4",
     "react": "18.2.0",
     "react-freeze": "1.0.3",
-    "react-i18next": "12.1.4",
+    "react-i18next": "12.1.5",
     "react-native": "0.71.8",
     "react-native-bootsplash": "4.5.0",
     "react-native-code-push": "7.1.0",
