+++ template/src/appInfrastructure/localisation/dictionaries/de/dev.json
@@ -1,24 +1,25 @@
 {
   "DevMode": {
-    "Hint_Header": "Hint",
-    "Hint_Text": "You can activate a \"DevMode\" by pressing 10 times on logo.\n\n<0>Password: </0><1>{{password}}</1>",
-    "Header": "Dev mode confirmation",
-    "Description": "You are going to open the Developer menu. If you know the password type it down. We are hope you are know what you do",
-    "PasswordInput": "Developer access password",
-    "Login": "Log in (Dev mode)",
-    "PasswordPlaceholder": "password",
-    "Logout": "Logout dev mode",
+    "Hint_Header": "Hinweis",
+    "Hint_Text": "Sie können einen \"DevMode\" aktivieren, indem Sie zehnmal auf das Logo drücken.\n\n<0>Passwort: </0><1>{{password}}</1>",
+    "Header": "Bestätigung des Entwicklungsmodus",
+    "Beschreibung": "Sie werden das Entwicklermenü öffnen. Wenn Sie das Passwort kennen, geben Sie es ein. Wir hoffen, dass Sie wissen, was Sie tun",
+    "PasswordInput": "Entwickler-Zugriffspasswort",
+    "Login": "Anmelden (Dev-Modus)",
+    "PasswordPlaceholder": "Passwort",
+    "Logout": "Entwicklermodus abmelden",
     "User": {
-      "Header": "User info",
-      "Logout": "Sign out",
+      "Header": "Benutzerinformationen",
+      "Logout": "Abmelden",
       "UserId": "<0>UserId:</0> {{value}}"
     },
     "App": {
-      "Header": "App info",
-      "Language": "<0>The interface language:</0> Deutsch",
-      "Env": "<0>Environment:</0> {{value}}",
-      "AppVersion": "<0>Build Version:</0> {{value}}",
-      "CodePush": "<0>CodePush Version:</0> {{value}}",
+      "Header": "App-Info",
+      "Language": "<0>Die Sprache der Benutzeroberfläche:</0> Deutsch",
+      "Env": "<0>Umgebung:</0> {{value}}",
+      "Server": "<0>Server-URL:</0> {{value}}",
+      "AppVersion": "<0>Build-Version:</0> {{value}}",
+      "CodePush": "<0>CodePush-Version:</0> {{value}}",
       "DevMode": "<0>Dev Mode:</0> {{value}}",
       "Hermes": "<0>Hermes:</0>  {{value}}",
       "HermesVersion": "<0>Hermes Version:</0>  {{value}}"
@@ -40,11 +41,11 @@
       "OpenSwipe": "Open swipe modal"
     },
     "Errors": {
-      "Header": "Errors tests",
-      "RequestWithToast": "Invalid request (toast)",
-      "Request": "Invalid request",
-      "JsCrash": "JS Crash",
-      "NativeCrash": "Native Crash"
+      "Header": "Fehlertests",
+      "RequestWithToast": "Ungültige Anfrage (Toast)",
+      "Request": "Ungültige Anfrage",
+      "JsCrash": "JS-Absturz",
+      "NativeCrash": "Native Absturz"
     },
     "Network": {
       "Header": "Network state info",
