+++ template/src/appInfrastructure/localisation/dictionaries/localization.en.json
@@ -1,0 +1,91 @@
+{
+  "Common_App": "RN Template App",
+  "Common_yes": "yes",
+  "Common_no": "no",
+  "Common_cancel": "cancel",
+  "Common_loading": "Loading...",
+  "Common_logout": "Log out",
+  "Common_back": "Back",
+  "Language_en": "English",
+  "Language_de": "Deutsch",
+  "Language_ru": "Russian",
+  "LoginScreen": {
+    "Login": "Login",
+    "LoginPlaceholder": "Write login",
+    "Code": "Code",
+    "CodePlaceholder": "Write access code",
+    "Header": "Log In",
+    "Text": "Click on button to simulate auth.\nYou can type any credentials - it is fake",
+    "Button": "Log in"
+  },
+  "ErrorBoundary_Text": "The app run into a problem and could not continue. We apologise for any inconvenience that has caused!\nPress the button below to restart the app.\nHopefully we have already known that it happened and trying to fix it it.",
+  "ErrorBoundary_Button": "Re-launch",
+  "ErrorBoundary_SendReport": "Send to bug tracker",
+  "ErrorBoundary_Header": "Oops,\nSomething went wrong...",
+  "NetworkErrorBoundary_Text": "We couldn't connect to our server. Maybe yor internet connection is down?",
+  "NetworkErrorBoundary_Button": "Refresh",
+  "NetworkErrorBoundary_Header": "Network problems",
+  "Home": {
+    "Greeting": "Hello Bro!",
+    "Description": "It is a dashboard page",
+    "Libs": {
+      "Header": "Packages",
+      "ReactAndRedux_Header": "React-Redux",
+      "ReactAndRedux_Text": "- React-Native\n- Redux\n- Redux-persist\n- react-redux",
+      "Network_Header": "Network",
+      "Network_Text": "- Axios\n- React-native-netinfo\n- React-native-restart\n- react-query\n- react-query-nswag",
+      "UI_Header": "UI",
+      "UI_Text": "- React navigation\n- React-native-modals\n- React-native-bootsplash\n- SafeAreaContext\n- KeyboardAware",
+      "Services_Header": "Services",
+      "Services_Text": "- Sentry\n- Firebase\n- CodePush\n- Matomo",
+      "Utils_Header": "Utility",
+      "Utils_Text": "- i18next\n- Date-fns\n- React-native-config (ENV)\n- React-native-startup-time\n- React-native-make  (helper for app icons)\n- Why-did-you-render"
+    },
+    "Hooks": {
+      "Header": "Hooks",
+      "Text": "We also have extra custom hooks. To see them click on button upper. Use them to write less code",
+      "Description": "We have got such hooks as:\n\nuseNotifications - initialize and handle pushes\nuseCustomBackBehaviour - to override system back button\nuseModal - add modal management (We've just open IT with this hook)\nuseNetworkError - handle network connection and redirect to error screen\nuseStartupTime - measure app launch time since native to js loading moment",
+      "Button": "Take a look",
+      "Modal": "Hooks modal"
+    },
+    "Authorized": {
+      "Header": "Backend communication",
+      "Text": "You can login and work without our <0>Backend</0> template (<1>{{url}}<1/>)",
+      "Button": "Go next"
+    }
+  },
+  "Language": "The interface language is English",
+  "Header_Language": "Language",
+  "App_Version": "App version: {{version}}",
+  "DevMode": {
+    "Hint_Header": "Hint",
+    "Hint_Text": "You can activate a \"DevMode\" by pressing 10 times on logo.\n\n<0>Password: </0><1>{{password}}</1>",
+    "Header": "Dev mode confirmation",
+    "Description": "You are going to open the Developer menu. If you know the password type it down. We are hope you are know what you do",
+    "PasswordInput": "Developer access password",
+    "Login": "Log in (Dev mode)",
+    "PasswordPlaceholder": "password",
+    "Logout": "Logout dev mode",
+    "User": {
+      "Header": "User info",
+      "Logout": "Sign out",
+      "UserId": "<0>UserId:</0> {{value}}"
+    },
+    "App": {
+      "Header": "App info",
+      "Language": "<0>The interface language:</0> English",
+      "Env": "<0>Environment:</0> {{value}}",
+      "Server": "<0>Server URL:</0> {{value}}",
+      "AppVersion": "<0>Build Version:</0> {{value}}",
+      "CodePush": "<0>CodePush Version:</0> {{value}}",
+      "DevMode": "<0>Dev Mode:</0> {{value}}"
+    },
+    "Errors": {
+      "Header": "Errors tests",
+      "RequestWithToast": "Invalid request (toast)",
+      "Request": "Invalid request",
+      "JsCrash": "JS Crash",
+      "NativeCrash": "Native Crash"
+    }
+  }
+}
