+++ template/src/appInfrastructure/localisation/dictionaries/localization.ru.json
@@ -1,0 +1,28 @@
+{
+  "Common_App": "RN Template App",
+  "Common_yes": "да",
+  "Common_no": "нет",
+  "Common_cancel": "отмена",
+  "Common_loading": "Загрузка...",
+  "Common_logout": "Выход",
+  "Language_en": "Английский",
+  "Language_de": "Немецкий",
+  "Language_ru": "Русский",
+  "Login": {
+    "Header": "Авторизация",
+    "Text": "Нажмите на кнопку, чтобы имитировать авторизацию",
+    "Button": "Войти"
+  },
+  "ErrorBoundary_Text": "В приложение произошла ошибка. Мы просим прощения за неудобство!\nНажмите кнопку ниже, чтобы перезагрузить приложение.\nК счастью мы уже знаем что это произошло и пытаемся устранить проблему.",
+  "ErrorBoundary_Button": "Перезапустить",
+  "ErrorBoundary_Header": "Ой-ой-ой,\nЧто-то пошло не так...",
+  "NetworkErrorBoundary_Text": "Не удалось соединиться с сервером. Может у вас пробелмы с доступом к интернету?",
+  "NetworkErrorBoundary_Button": "Обнвоить",
+  "NetworkErrorBoundary_Header": "Проблемы интернет соединения",
+  "Home_Description": "Это шаблон с базовой функциональностью. В него входят такие библиотеки как: \n- Redux, redux-persist\n- React navigation + sub deps\n- React-native-netinfo\n- React-native-restart\n- Sentry\n- Axios + react-query + react-query-nswag\n- React-native-modals\n- React-native-startup-time\n- Firebase (push notifications + performance)\n- i18next\n- React-native-bootsplash\n- React-native-make  (helper for app icons)\n- CodePush\n- Why-did-you-render\nreact-native-config (ENV variables)",
+  "Home_Hooks_Text": "Так же были добавлены несколько кастомных хуков, чтоыб посмотреть их кликни на кнпоку ниже",
+  "Home_Hooks_Button": "Взглянуть",
+  "Home_Hooks_Description": "We have got such hooks as:\n\nuseNotifications - initialize and handle pushes\nuseCustomBackBehaviour - to override system back button\nuseModal - add modal management (We've just open IT with this hook)\nuseNetworkError - handle network connection\nuseStartupTime - measure app launch time since native to js loading moment",
+  "Language": "Язык интерфейса русский",
+  "App_Version": "Версия приложения: {{version}}"
+}
