+++ template/src/components/languageSwitcher/languageItem.component.tsx
@@ -1,10 +1,25 @@
 import React, { useCallback } from 'react';
-import { Image, Pressable, Text } from 'react-native';
-import { itemButtonStyles, itemIconStyles, itemLocalStyles } from './languageSwitcher.styles';
+import {Image, Pressable, StyleProp, Text} from 'react-native';
+import { itemButtonStyles, itemIconStyles, itemLocalStyles} from './languageSwitcher.styles';
 import { LanguageItemProps } from './languageSwitcher';
-import { getPressableStyles } from '../../commons/styles/styles';
 
 const checkIcon = require('assets/images/check_20.png');
+
+
+function getPressableStyles<T, K>(
+    pressed: boolean,
+    selected: boolean,
+    baseStyles: StyleProp<K>,
+    stateStyles: T
+): StyleProp<K> {
+  const containerStyle: StyleProp<K>[] = [baseStyles, {} as Record<any, any>];
+
+  const buttonStateSelector = pressed ? 'pushed' : selected ? 'enabled' : 'default';
+  // @ts-ignore
+  containerStyle.push(stateStyles[buttonStateSelector]);
+  return containerStyle;
+}
+
 
 export const LanguageItem = (props: LanguageItemProps) => {
   const getButtonStyle = useCallback(
